refactor: optimize tests slightly
